Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/workers/dedicated-worker-in-data-url-context.window-expected.txt

CONSOLE MESSAGE: SecurityError: The operation is insecure.

Harness Error (TIMEOUT), message = null

FAIL Create a dedicated worker in a data url frame assert_equals: expected "PASS" but got "Worker construction unexpectedly synchronously failed"
FAIL Create a dedicated worker in a data url dedicated worker promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
TIMEOUT Create a data url dedicated worker in a data url frame Test timed out
NOTRUN Create a data url dedicated worker in a data url dedicated worker
